Refactor: split QTree struct into a source file
